Separated error term from private key
In RLWE, a public/private key pair is really a public key/private key/error term triple. Originally, the error term was included as a variable in the private key object, but this has been removed. Once the public key has been established the error term is no longer needed so the error term is created during generation of the public key and discarded. The only detriment is that it is impossible to recreate the same public key from a given private key.